+++ code/programs/haskell/build-tool/test/Spec.hs
@@ -2,7 +2,7 @@ import Test.Hspec
 
 import BuildToolSpec (buildToolSpec)
 import HashingSpec (hashingSpec)
-import ResolutionUtf8Spec (resolutionCabalSpec, resolutionGoSpec, resolutionPerlSpec, resolutionPythonSpec, resolutionRubySpec, resolutionRustSpec, resolutionSwiftSpec, resolutionUtf8Spec)
+import ResolutionUtf8Spec (resolutionCabalSpec, resolutionElixirSpec, resolutionGoSpec, resolutionPerlSpec, resolutionPythonSpec, resolutionRubySpec, resolutionRustSpec, resolutionSwiftSpec, resolutionUtf8Spec)
 
 main :: IO ()
 main = hspec spec
@@ -14,6 +14,7 @@ spec = do
     resolutionUtf8Spec
     resolutionCabalSpec
     resolutionGoSpec
+    resolutionElixirSpec
     resolutionPythonSpec
     resolutionRustSpec
     resolutionRubySpec